requirements.txt
flask==2.3.2
openai==0.27.8
nltk==3.8.1
python-dotenv==1.0.0